pyproject.toml
[project]
name = "generic-decorator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
]
